requirements.txt
Django==1.11.4
Pillow==4.1.1
mysqlclient==1.3.10
django-cleanup==0.4.2
django-summernote==0.8.7
python-memcached==1.58
django-filter==1.0.4
django-constance==2.0.0
django-picklefield==0.3.2